LayoutTests/platform/gtk/css-dark-mode/default-colors-expected.txt


PASS Body text color is black, since the page isn't in dark mode 
FAIL View base background color is white, since the page isn't in dark mode assert_equals: expected "rgb(255, 255, 255)" but got "rgb(246, 245, 244)"
PASS Set dark appearance 
FAIL Body text color is black, since the page hasn't set a supported color scheme assert_equals: expected "rgb(0, 0, 0)" but got "rgb(255, 255, 255)"
FAIL View base background color is white, since the page hasn't set a supported color scheme assert_equals: expected "rgb(255, 255, 255)" but got "rgb(246, 245, 244)"
PASS Set prefers-color-schemes: light dark on the body 
PASS Body text color is white 
FAIL View base background color is a dark grey assert_equals: expected "rgb(30, 30, 30)" but got "rgb(53, 53, 53)"
